pyproject.toml
[project]
name = "tennet-afrr-archive"
version = "0.1.0"
description = "Downloading and storing the TenneT AFRR bid ladders at regular intervals "
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deltalake>=1.4.2",
    "polars>=1.0.0",
    "python-dotenv>=1.2.1",
    "tenneteu-py>=0.1.5",
]
